requirements.txt
google-generativeai
google-cloud-bigquery
python-dotenv
pdfplumber
requests